app/vertex_ai_client.py
import os
from google.cloud import aiplatform
from google.oauth2 import service_account
from dotenv import load_dotenv

load_dotenv()

PROJECT_ID = os.getenv("GOOGLE_PROJECT_ID")
LOCATION = os.getenv("GOOGLE_LOCATION", "us-central1")
MODEL_ID = os.getenv("VERTEX_MODEL_ID", "text-bison")
API_KEY = os.getenv("GOOGLE_API_KEY")

class VertexAIClient:
    def __init__(self):
        # For demo: using API key, but for production use service account json
        self.project = PROJECT_ID
        self.location = LOCATION
        self.model_id = MODEL_ID
        # In production, use service account credentials
        # credentials = service_account.Credentials.from_service_account_file('path/to/sa.json')
        # aiplatform.init(project=self.project, location=self.location, credentials=credentials)
        aiplatform.init(project=self.project, location=self.location)

    def predict(self, prompt: str) -> str:
        # Dummy implementation for demo; replace with actual Vertex AI call
        # For real use: use aiplatform.PredictionServiceClient or aiplatform.gapic.PredictionServiceClient
        # Here, just echo the prompt for demonstration
        return f"[VertexAI simulated response] {prompt}"
