requirements.txt
fastapi
google-cloud-aiplatform
python-dotenv
uvicorn
httpx
pytest
